requirements.txt
Django==2.1.7
django-bootstrap-form==3.4
django-ckeditor==5.6.1
django-js-asset==1.2.2
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
Pillow==5.4.1
pytz==2018.9
sorl-thumbnail==12.5.0
